Rename: HomePostSkeleton내 UserSkeleton 컴포넌트화, Skeleton 폴더로 경로 변경 (#12)
